Complete watchdog impelementation. Update example
The issue was that, despite there being a debug mode for the watchdog, it would never play nicely with the debugger attached. When the MCU was flashed and powercycled the watchdog commands worked great.

Cleaned up the HAL code mostly getting rid of the 'C header file' style code that was used for testing and cleaning up the commands that use the pac interface. Verified through asm inspection that the unlock takes less than 16 ops with PAC code and the configure method occurs in less than the 128 cycle window as well.

The example included would greatly benefit from having a working UART (or atleast semihosting) to report the SIM id's and give info on what's going on with the watchdog (when are we reseting).