Finalize Analog implementation. Update examples and docs. Fix AdcResolution. Activate Bandgap at Adc Enable.
Analog impl moved back to adc module. Channel is now implemented on the Analog type. On-chip channel types are effectively dummy pins so that the same interface can be utilized. Pins are wrapped by the analog type so they can be returned in the same typestate they were received. The into_analog and outof_analog impls were done in a macro instead of relying on the the Channel trait method.

Changed the gpio module's Into impls to From impls. Mostly to make clippy happy.